remove zip utils and place everything in multimediazipdownload
